+++ core/src/main/java/cucumber/api/DataTable.java
@@ -6,13 +6,11 @@ import cucumber.runtime.table.DiffableRow;
 import cucumber.runtime.table.TableConverter;
 import cucumber.runtime.table.TableDiffException;
 import cucumber.runtime.table.TableDiffer;
-import cucumber.runtime.table.TypeReference;
 import cucumber.runtime.xstream.LocalizedXStreams;
 import gherkin.formatter.PrettyFormatter;
 import gherkin.formatter.model.DataTableRow;
 import gherkin.formatter.model.Row;
 
-import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -76,20 +74,10 @@ public class DataTable {
     }
 
     /**
-     * Converts the table to a 2D array.
-     *
      * @return a List of List of String.
      */
     public List<List<String>> raw() {
         return this.raw;
-    }
-
-    public <T> T convert(Type type) {
-        return tableConverter.convert(type, this, false);
-    }
-
-    public <T> T convert(Type type, boolean transposed) {
-        return tableConverter.convert(type, this, transposed);
     }
 
     /**
@@ -98,34 +86,45 @@ public class DataTable {
      *
      * @return a List of Map.
      */
-    public <K, V> List<Map<K, V>> asMaps() {
-        return asList(new TypeReference<Map<String, String>>() {
-        }.getType());
+    public <K, V> List<Map<K, V>> asMaps(Class<K> keyType, Class<V> valueType) {
+        return tableConverter.toMaps(this, keyType, valueType);
     }
 
     /**
      * Converts the table to a single Map. The left column is used as keys, the right column as values.
      *
      * @return a Map.
+     * @throws cucumber.runtime.CucumberException if the table doesn't have 2 columns.
      */
-    public <K, V> Map<K, V> asMap() {
-        return tableConverter.convert(new TypeReference<Map<String, String>>() {
-        }.getType(), this, false);
+    public <K, V> Map<K, V> asMap(Class<K> keyType, Class<V> valueType) {
+        return tableConverter.toMap(this, keyType, valueType);
     }
 
     /**
-     * Converts the table to a List of objects. The top row is used to identifies the fields/properties
-     * of the objects.
-     * <p/>
-     * Backends that support generic types can declare a parameter as a List of a type, and Cucumber will
-     * do the conversion automatically.
+     * Converts the table to a List.
      *
-     * @param type the type of the result (should be a {@link List} generic type)
-     * @param <T>  the type of each object
-     * @return a list of objects
+     * If {@code itemType} is a scalar type the table is flattened.
+     *
+     * Otherwise, the top row is used to name the fields/properties and the remaining
+     * rows are turned into list items.
+     *
+     * @param itemType the type of the list items
+     * @param <T>      the type of the list items
+     * @return a List of objects
      */
-    public <T> List<T> asList(Type type) {
-        return tableConverter.toList(type, this, false);
+    public <T> List<T> asList(Class<T> itemType) {
+        return tableConverter.toList(this, itemType);
+    }
+
+    /**
+     * Converts the table to a List of List of scalar.
+     *
+     * @param itemType the type of the list items
+     * @param <T>      the type of the list items
+     * @return a List of List of objects
+     */
+    public <T> List<List<T>> asLists(Class<T> itemType) {
+        return tableConverter.toLists(this, itemType);
     }
 
     public List<String> topCells() {
@@ -199,16 +198,6 @@ public class DataTable {
 
     public TableConverter getTableConverter() {
         return tableConverter;
-    }
-
-    public List<String> flatten() {
-        List<String> result = new ArrayList<String>();
-        for (List<String> rows : raw()) {
-            for (String cell : rows) {
-                result.add(cell);
-            }
-        }
-        return result;
     }
 
     public DataTable transpose() {