+++ core/src/test/java/cucumber/runtime/table/ToDataTableTest.java
@@ -7,7 +7,6 @@ import cucumber.runtime.xstream.LocalizedXStreams;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.lang.reflect.Type;
 import java.util.Date;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -31,7 +30,7 @@ public class ToDataTableTest {
 
     @Test
     public void converts_list_of_beans_to_table() {
-        List<UserPojo> users = tc.toList(UserPojo.class, personTable());
+        List<UserPojo> users = tc.toList(personTable(), UserPojo.class);
         DataTable table = tc.toTable(users);
         assertEquals("" +
                 "      | credits | name        | birthDate  |\n" +
@@ -42,7 +41,7 @@ public class ToDataTableTest {
 
     @Test
     public void converts_list_of_beans_with_null_to_table() {
-        List<UserPojo> users = tc.toList(UserPojo.class, personTableWithNull());
+        List<UserPojo> users = tc.toList(personTableWithNull(), UserPojo.class);
         DataTable table = tc.toTable(users, "name", "birthDate", "credits");
         assertEquals("" +
                 "      | name        | birthDate  | credits |\n" +
@@ -54,12 +53,12 @@ public class ToDataTableTest {
     @Test
     public void gives_a_nice_error_message_when_field_is_missing() {
         try {
-            tc.toList(UserPojo.class, TableParser.parse("" +
+            tc.toList(TableParser.parse("" +
                     "| name        | birthDate  | crapola  |\n" +
                     "| Sid Vicious | 10/05/1957 | 1,000    |\n" +
                     "| Frank Zappa | 21/12/1940 | 3,000    |\n" +
-                    "", PARAMETER_INFO)
-            );
+                    "", PARAMETER_INFO),
+                    UserPojo.class);
             fail();
         } catch (CucumberException e) {
             assertEquals("No such field cucumber.runtime.table.ToDataTableTest$UserPojo.crapola", e.getMessage());
@@ -69,11 +68,12 @@ public class ToDataTableTest {
     @Test
     public void gives_a_nice_error_message_when_primitive_field_is_null() {
         try {
-            tc.toList(PojoWithInt.class, TableParser.parse("" +
+            tc.toList(TableParser.parse("" +
                     "| credits     |\n" +
                     "| 5           |\n" +
                     "|             |\n" +
-                    "", PARAMETER_INFO)
+                    "", PARAMETER_INFO),
+                    PojoWithInt.class
             );
             fail();
         } catch (CucumberException e) {
@@ -84,10 +84,11 @@ public class ToDataTableTest {
     @Test
     public void gives_a_meaningfull_error_message_when_field_is_repeated() {
         try {
-            tc.toList(UserPojo.class, TableParser.parse("" +
+            tc.toList(TableParser.parse("" +
                     "| credits     | credits     |\n" +
                     "| 5           | 5           |\n" +
-                    "", PARAMETER_INFO)
+                    "", PARAMETER_INFO),
+                    UserPojo.class
             );
             fail();
         } catch (CucumberException e) {
@@ -97,7 +98,7 @@ public class ToDataTableTest {
 
     @Test
     public void converts_list_of_beans_to_table_with_explicit_columns() {
-        List<UserPojo> users = tc.toList(UserPojo.class, personTable());
+        List<UserPojo> users = tc.toList(personTable(), UserPojo.class);
         DataTable table = tc.toTable(users, "name", "birthDate", "credits");
         assertEquals("" +
                 "      | name        | birthDate  | credits |\n" +
@@ -108,7 +109,7 @@ public class ToDataTableTest {
 
     @Test
     public void diffs_round_trip() {
-        List<UserPojo> users = tc.toList(UserPojo.class, personTable());
+        List<UserPojo> users = tc.toList(personTable(), UserPojo.class);
         personTable().diff(users);
     }
 
@@ -136,9 +137,7 @@ public class ToDataTableTest {
                 "      | 0.5 | 1.5     |\n" +
                 "      | 99  | 1,000.5 |\n" +
                 "", table.toString());
-        Type listOfDoubleType = new TypeReference<List<Double>>() {
-        }.getType();
-        List<List<Double>> actual = tc.toList(listOfDoubleType, table);
+        List<List<Double>> actual = tc.toLists(table, Double.class);
         assertEquals(lists, actual);
     }