+++ scala/src/test/scala/cucumber/runtime/scala/test/StepDefs.scala
@@ -6,6 +6,7 @@ import cucumber.api.DataTable
 import junit.framework.Assert._
 import scala.collection.JavaConversions._
 import cucumber.runtime.scala.model.{Snake, Person}
+import java.util
 
 /**
  * Test step definitions to exercise Scala cucumber
@@ -19,7 +20,8 @@ class CukesStepDefinitions extends ScalaDsl with EN {
   var calorieCount = 0.0
 
   Given("""^I have the following foods :$"""){ (table:DataTable) =>
-    calorieCount = table.asMaps().map(_.get("CALORIES")).map(_.toDouble).fold(0.0)(_+_)
+    val maps: util.List[util.Map[String, String]] = table.asMaps(classOf[String], classOf[String])
+    calorieCount = maps.map(_.get("CALORIES")).map(_.toDouble).fold(0.0)(_+_)
   }
   And("""^have eaten (.*) calories today""") { (calories:Double) =>
     assertEquals(calories, calorieCount)
@@ -125,7 +127,7 @@ class CukesStepDefinitions extends ScalaDsl with EN {
   }
 
   Given("""^I have a table the sum of all rows should be (\d+) :$"""){ (value:Int, table:DataTable) =>
-    assertEquals(value, table.flatten.drop(1).map(_.toInt).foldLeft(0)(_+_))
+    assertEquals(value, table.asList(classOf[String]).drop(1).map(_.toInt).foldLeft(0)(_+_))
   }
 
   var snake:Snake = null